remove unnecessary cast
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@3435 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/export/JRHtmlExporter.java
@@ -2274,7 +2274,7 @@ public class JRHtmlExporter extends JRAbstractExporter
 
 	public String toSizeUnit(int size)
 	{
-		return String.valueOf((int)(zoom * (float)size)) + sizeUnit;
+		return String.valueOf((int)(zoom * size)) + sizeUnit;
 	}
